Fix numeric overflow in JSON integral coercion and add tests

+++ java/src/org/openqa/selenium/json/JsonInput.java
@@ -263,7 +263,7 @@ public class JsonInput implements Closeable {
         return Long.valueOf(builder.toString());
       }
 
-      return new BigDecimal(builder.toString());
+      return new BigDecimal(builder.toString()).doubleValue();
     } catch (NumberFormatException e) {
       throw new JsonException("Unable to parse to a number: " + builder + ". " + input, e);
     }
